+++ controllers/single_page/dashboard/store/products/groups.php
@@ -25,7 +25,7 @@ class Groups extends DashboardPageController
         $this->requireAsset('css', 'select2');
         $this->requireAsset('javascript', 'select2');
 
-        if ($this->post()) {
+        if ($this->post() && $this->token->validate('community_store')) {
             $errors = $this->validateGroup($this->post());
             $this->error = $errors;
             if (!$errors->has()) {
@@ -56,7 +56,7 @@ class Groups extends DashboardPageController
 
         $group = StoreGroup::getByID($gID);
 
-        if ($this->post()) {
+        if ($this->post() && $this->token->validate('community_store')) {
             $this->error = null; //clear errors
             $errors = $this->validateGroup($this->post());
             $this->error = $errors;
@@ -99,6 +99,8 @@ class Groups extends DashboardPageController
 
     public function deletegroup($gID)
     {
-        StoreGroup::getByID($gID)->delete();
+        if ($this->token->validate('community_store')) {
+            StoreGroup::getByID($gID)->delete();
+        }
     }
 }
